refactor(canister-core): dissolve stable_store into domain modules + store
`stable_store.rs` was a misleadingly-named grab-bag: it held the persisted data
*types* (which read like a "store" but carry no logic) next to their byte
encodings. Split those two concerns:

- Each type moves to its domain module: AssetMeta/EncodingMeta → asset.rs,
  TokenMeta/ProtectionSettings → protection.rs, RedirectRules → redirect.rs,
  ContentChunkKey/ChunkCert → blob_store.rs. Each domain module now owns its own
  data shape.
- Every `impl Storable` (the stable-memory byte encoding) + the impl_cbor_storable
  macro move to store.rs, so the storage layer owns *both* the memory layout (the
  MemoryId map) and the byte format — a format change is now contained to one
  module. The two encoding-only newtypes with no domain (AuthorizedSet, StateHash)
  are defined there too. The golden-vector encoding tests moved with them.

stable_store.rs is deleted. No behavior change — the byte formats are copied
verbatim; the Storable-roundtrip golden vectors and the upgrade-roundtrip test
still pass, pinning the on-stable format.

Verified: builds (incl. canbench-rs), 158 unit tests pass, fmt + clippy clean.
